Move Datastore from sdks/java/core to io/gcp
* Remove DataflowDatastoreIOTest. Dataflow runner translation of display
  data should be tested via code in the same module.
